Load original file when unable to load cached file.

+++ src/window.cpp
@@ -944,10 +944,13 @@ bool Window::addDocument(const QString& file, const QString& datafile, int posit
 	Document* document = new Document(file, m_current_wordcount, m_current_time, this);
 	m_documents->addDocument(document);
 	document->loadTheme(m_sessions->current()->theme());
-	document->loadFile(path, m_save_positions ? position : -1);
-	if (datafile != file) {
-		document->text()->document()->setModified(!compareFiles(file, datafile));
-		QFile::remove(datafile);
+	if (document->loadFile(path, m_save_positions ? position : -1)) {
+		if (datafile != file) {
+			document->text()->document()->setModified(!compareFiles(file, datafile));
+			QFile::remove(datafile);
+		}
+	} else {
+		document->loadFile(file, m_save_positions ? position : -1);
 	}
 	connect(document, SIGNAL(changed()), this, SLOT(updateDetails()));
 	connect(document, SIGNAL(changed()), this, SLOT(updateProgress()));
